Player, Soldier, Character
The following is a Dependency-Injection based implementation of Character, Player, and Soldier.

Soldier is an implementation of Character. CharacterSpawner is responsible for instantiating the Soldier object and supplying it to the Player. Player runs the input loops and sends its inputs to the Character.

If you click on the screen, a CharacterSpawnerTest object will run this injector and spawn a Soldier. This can then be controlled by WASD (or similar controls).